Refactor init command to initializer package
Moved the `init` command implementation from `cmd/init.go` to `internal/initializer/cmd.go`. This improves code organization by grouping initialization-related logic within the `initializer` package. Adjusted imports and references accordingly in `main.go`.
